don't remove our distinguished comments

+++ src/reddit/mod.rs
@@ -9,7 +9,7 @@ use anyhow::{anyhow, bail, Context};
 use config::{SubredditModerateConfig, SubredditsConfig};
 use flairs::{FlairChangeConfig, PostFlairCache, SubredditFlairConfig};
 use roux::{
-    api::ThingFullname,
+    api::{Distinguished, ThingFullname},
     client::{OAuthClient, RedditClient as RouxRedditClient},
     models::Distinguish,
     util::RouxError,
@@ -28,8 +28,8 @@ use crate::{
     webhook::{
         create_deleted_downvoted_comment, create_detection_message,
         create_error_processing_message, create_error_processing_post,
-        create_generic_error_message, create_inbox_message, create_multiple_error_message,
-        Message as DiscordMessage, WebhookClient,
+        create_generic_error_message, create_inbox_message, create_moderator_downvoted_comment,
+        create_multiple_error_message, Message as DiscordMessage, WebhookClient,
     },
     RedditInfo,
 };
@@ -548,16 +548,29 @@ impl<'a> RedditClient<'a> {
 
         for comment in comments {
             if !comment.score_hidden() && comment.score() < 0 {
-                println!(
-                    "Removing downvoted {:?} on {:?} by /u/{}",
-                    comment.name(),
-                    comment.link_title(),
-                    comment.link_author()
-                );
-                comment.delete()?;
-                if let Some(webhook) = &mut self.webhook {
-                    let message = create_deleted_downvoted_comment(&comment);
-                    webhook.send(&message)?;
+                if matches!(comment.distinguished(), Distinguished::None) {
+                    println!(
+                        "Removing downvoted {:?} on {:?} by /u/{}",
+                        comment.name(),
+                        comment.link_title(),
+                        comment.link_author()
+                    );
+                    comment.delete()?;
+                    if let Some(webhook) = &mut self.webhook {
+                        let message = create_deleted_downvoted_comment(&comment);
+                        webhook.send(&message)?;
+                    }
+                } else {
+                    println!(
+                        "NOT removing downvoted {:?} on {:?} by /u/{}",
+                        comment.name(),
+                        comment.link_title(),
+                        comment.link_author()
+                    );
+                    if let Some(webhook) = &mut self.webhook {
+                        let message = create_moderator_downvoted_comment(&comment);
+                        webhook.send(&message)?;
+                    }
                 }
             }
         }
